move cache test to datasource
